From Xu Lu and Xuguang Wang (OU): Add the FGAT and HAFS ensemble capabilities in HAFS DA workflow. POC: xuguang.wang@ou.edu

+++ scripts/exhafs_analysis.sh
@@ -12,6 +12,9 @@ export use_bufr_nr=${use_bufr_nr:-no}
 export RUN_GSI=${RUN_GSI:-NO}
 export RUN_GSI_VR=${RUN_GSI_VR:-NO}
 export hybrid_3denvar_gdas=${hybrid_3denvar_gdas:-yes}
+export RUN_HAFS_ENS=${RUN_HAFS_ENS:-NO}
+export G_RATIO_ENS=${G_RATIO_ENS:-1}
+export FGAT_OPT=${FGAT_OPT:-NO} 
 
 TOTAL_TASKS=${TOTAL_TASKS:-2016}
 NCTSK=${NCTSK:-12}
@@ -33,6 +36,13 @@ mnprior=`echo ${CDATEprior} | cut -c5-6`
 dyprior=`echo ${CDATEprior} | cut -c7-8`
 hhprior=`echo ${CDATEprior} | cut -c9-10`
 PDYprior=`echo ${CDATEprior} | cut -c1-8`
+
+if [ ${FGAT_OPT} = YES ]; then 
+ CDATEpri=`${NDATE} -3 $CDATE`
+ cyc_pri=`echo ${CDATEpri} | cut -c9-10`
+ CDATEaft=`${NDATE} +3 $CDATE`
+ cyc_aft=`echo ${CDATEaft} | cut -c9-10`
+fi
 
 export COMhafsprior=${COMhafsprior:-${COMhafs}/../../${CDATEprior}/${STORMID}}
 export WORKhafsprior=${WORKhafsprior:-${WORKhafs}/../../${CDATEprior}/${STORMID}}
@@ -61,7 +71,35 @@ fi
 ${NCP} ${RESTARTinp}/oro_data.nc ./fv3_oro_data
 ${NCP} ${RESTARTinp}/atmos_static.nc ./fv3_atmos_static
 ${NCP} ${RESTARTinp}/grid_spec.nc ./fv3_grid_spec
-
+if [ ${FGAT_OPT} = YES ]; then 
+ if [ ${RUN_GSI_VR} = "YES" ]; then
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_pri}0000.coupler.res ./coupler.res_03
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_pri}0000.fv_core.res.nc ./fv3_akbk_03
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_pri}0000.sfc_data.nc ./fv3_sfcdata_03
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_pri}0000.fv_srf_wnd.res.tile1.nc ./fv3_srfwnd_03
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_pri}0000.fv_core.res.tile1.nc ./fv3_dynvars_03
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_pri}0000.fv_tracer.res.tile1.nc ./fv3_tracer_03
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_aft}0000.coupler.res ./coupler.res_09
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_aft}0000.fv_core.res.nc ./fv3_akbk_09
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_aft}0000.sfc_data.nc ./fv3_sfcdata_09
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_aft}0000.fv_srf_wnd.res.tile1.nc ./fv3_srfwnd_09
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_aft}0000.fv_core.res.tile1.nc ./fv3_dynvars_09
+  ${NCP} ${RESTARTinp}/${PDY}.${cyc_aft}0000.fv_tracer.res.tile1.nc ./fv3_tracer_09
+ else
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_pri}0000.coupler.res ./coupler.res_03
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_pri}0000.fv_core.res.nc ./fv3_akbk_03
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_pri}0000.sfc_data.nc ./fv3_sfcdata_03
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_pri}0000.fv_srf_wnd.res.tile1.nc ./fv3_srfwnd_03
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_pri}0000.fv_core.res.tile1.nc ./fv3_dynvars_03
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_pri}0000.fv_tracer.res.tile1.nc ./fv3_tracer_03
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_aft}0000.coupler.res ./coupler.res_09
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_aft}0000.fv_core.res.nc ./fv3_akbk_09
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_aft}0000.sfc_data.nc ./fv3_sfcdata_09
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_aft}0000.fv_srf_wnd.res.tile1.nc ./fv3_srfwnd_09
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_aft}0000.fv_core.res.tile1.nc ./fv3_dynvars_09
+  ${NCP} ${COMhafsprior}/RESTART/${PDY}.${cyc_aft}0000.fv_tracer.res.tile1.nc ./fv3_tracer_09
+ fi
+fi
 ${NCP} ${RESTARTinp}/${PDY}.${cyc}0000.coupler.res ./coupler.res
 ${NCP} ${RESTARTinp}/${PDY}.${cyc}0000.fv_core.res.nc ./fv3_akbk
 ${NCP} ${RESTARTinp}/${PDY}.${cyc}0000.sfc_data.nc ./fv3_sfcdata
@@ -72,13 +110,47 @@ ${NCP} ${RESTARTinp}/${PDY}.${cyc}0000.fv_tracer.res.tile1.nc ./fv3_tracer
 if [ $hybrid_3denvar_gdas = yes ]; then
 
 export L_HYB_ENS=.true.
-export N_ENS=80
+if [ $RUN_HAFS_ENS = YES ]; then
+  export N_ENS=${ENS_SIZE:-2}
+  export G_RATIO=${G_RATIO_ENS}  
+  export REO_ENS=5
+  for mem in $(seq -f '%03g' 1 ${N_ENS})
+  do
+   if  [ ${RUN_GSI_VR} = "YES" ]; then
+    ${NLN} ${COMhafs}/RESTART_analysis_vr_ens/member${mem}/${PDY}.${cyc}0000.coupler.res ./fv3SAR06_ens_mem${mem}-coupler.res
+    ${NLN} ${COMhafs}/RESTART_analysis_vr_ens/member${mem}/${PDY}.${cyc}0000.fv_core.res.nc ./fv3SAR06_ens_mem${mem}-fv3_akbk
+    ${NLN} ${COMhafs}/RESTART_analysis_vr_ens/member${mem}/${PDY}.${cyc}0000.sfc_data.nc ./fv3SAR06_ens_mem${mem}-fv3_sfcdata
+    ${NLN} ${COMhafs}/RESTART_analysis_vr_ens/member${mem}/${PDY}.${cyc}0000.fv_srf_wnd.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_srfwnd
+    ${NLN} ${COMhafs}/RESTART_analysis_vr_ens/member${mem}/${PDY}.${cyc}0000.fv_core.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_dynvars
+    ${NLN} ${COMhafs}/RESTART_analysis_vr_ens/member${mem}/${PDY}.${cyc}0000.fv_tracer.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_tracer
+   else
+    if [ -s ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/${PDY}.${cyc}0000.fv_core.res.tile1.nc ];then
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/${PDY}.${cyc}0000.coupler.res ./fv3SAR06_ens_mem${mem}-coupler.res
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/${PDY}.${cyc}0000.fv_core.res.nc ./fv3SAR06_ens_mem${mem}-fv3_akbk
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/${PDY}.${cyc}0000.sfc_data.nc ./fv3SAR06_ens_mem${mem}-fv3_sfcdata
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/${PDY}.${cyc}0000.fv_srf_wnd.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_srfwnd
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/${PDY}.${cyc}0000.fv_core.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_dynvars
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/${PDY}.${cyc}0000.fv_tracer.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_tracer
+    else
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/coupler.res ./fv3SAR06_ens_mem${mem}-coupler.res
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/fv_core.res.nc ./fv3SAR06_ens_mem${mem}-fv3_akbk
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/sfc_data.nc ./fv3SAR06_ens_mem${mem}-fv3_sfcdata
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/fv_srf_wnd.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_srfwnd
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/fv_core.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_dynvars
+     ${NLN} ${COMhafsprior}/RESTART_ENS/member${mem}/fv_tracer.res.tile1.nc ./fv3SAR06_ens_mem${mem}-fv3_tracer
+    fi
+   fi
+  done
+else
+  export N_ENS=80
+  export G_RATIO=1
+  export REO_ENS=1
 # Link ensemble members
-mkdir -p ensemble_data
-ENKF_SUFFIX="s"
-GSUFFIX=${GSUFFIX:-.nemsio}
-fhrs="06"
-for fhh in $fhrs; do
+  mkdir -p ensemble_data
+  ENKF_SUFFIX="s"
+  GSUFFIX=${GSUFFIX:-.nemsio}
+  fhrs="06"
+  for fhh in $fhrs; do
   rm -f filelist${fhh}
   for mem in $(seq -f '%03g' 1 ${N_ENS}); do
     if [ -s ${COMgfs}/enkfgdas.${PDYprior}/${hhprior}/mem${mem}/gdas.t${hhprior}z.atmf0${fhh}s${GSUFFIX:-.nemsio} ]; then
@@ -87,9 +159,9 @@ for fhh in $fhrs; do
       ${NLN} ${COMgfs}/enkfgdas.${PDYprior}/${hhprior}/mem${mem}/gdas.t${hhprior}z.atmf0${fhh}${GSUFFIX:-.nemsio} ./ensemble_data/enkfgdas.${PDYprior}${hhprior}.atmf0${fhh}_ens_${mem}
     fi
     echo "./ensemble_data/enkfgdas.${PDYprior}${hhprior}.atmf0${fhh}_ens_${mem}" >> filelist${fhh}
+   done
   done
-done
-
+fi
 fi	
 
 #---------------------------------------------- 
@@ -305,6 +377,8 @@ ${NCP} ${PARMgsi}/gsiparm.anl.tmp ./
 
 sed -e "s/_L_HYB_ENS_/${L_HYB_ENS:-.false.}/g" \
     -e "s/_N_ENS_/${N_ENS:-80}/g" \
+    -e "s/_G_RATIO_/${G_RATIO:-1}/g" \
+    -e "s/_REO_ENS_/${REO_ENS:-1}/g" \
     -e "s/_GRID_RATIO_FV3_REGIONAL_/${refine_ratio:-4}/g" \
     gsiparm.anl.tmp > gsiparm.anl
